Split out part of Utils to StringUtils
