+++ lib/kennel.rb
@@ -5,6 +5,7 @@ require "zeitwerk"
 require "English"
 
 require "kennel/version"
+require "kennel/string_utils"
 require "kennel/utils"
 require "kennel/progress"
 require "kennel/filter"
