+++ lib/kennel/utils.rb
@@ -15,29 +15,6 @@ module Kennel
     end
 
     class << self
-      def snake_case(string)
-        string
-          .gsub(/::/, "_") # Foo::Bar -> foo_bar
-          .gsub(/([A-Z]+)([A-Z][a-z])/, '\1_\2') # FOOBar -> foo_bar
-          .gsub(/([a-z\d])([A-Z])/, '\1_\2') # fooBar -> foo_bar
-          .tr("-", "_") # foo-bar -> foo_bar
-          .downcase
-      end
-
-      # for child projects, not used internally
-      def title_case(string)
-        string.split(/[\s_]/).map(&:capitalize) * " "
-      end
-
-      # simplified version of https://apidock.com/rails/ActiveSupport/Inflector/parameterize
-      def parameterize(string)
-        string
-          .downcase
-          .gsub(/[^a-z0-9\-_]+/, "-") # remove unsupported
-          .gsub(/-{2,}/, "-") # remove duplicates
-          .gsub(/^-|-$/, "") # remove leading/trailing
-      end
-
       def presence(value)
         value.nil? || value.empty? ? nil : value
       end
@@ -56,12 +33,6 @@ module Kennel
         return text unless force || Kennel.out.tty?
 
         "\e[#{COLORS.fetch(color)}m#{text}\e[0m"
-      end
-
-      def truncate_lines(text, to:, warning:)
-        lines = text.split(/\n/, to + 1)
-        lines[-1] = warning if lines.size > to
-        lines.join("\n")
       end
 
       def capture_stdout
@@ -127,10 +98,6 @@ module Kennel
         end
         workers.each(&:join)
         done.each { |d| raise d if d.is_a?(Exception) }
-      end
-
-      def natural_order(name)
-        name.split(/(\d+)/).each_with_index.map { |x, i| i.odd? ? x.to_i : x }
       end
 
       def retry(*errors, times:)
